Move tasks and configuration to own files
