+++ tasks/i18n.js
@@ -1,6 +1,6 @@
 module.exports = function(grunt) {
 
-  grunt.registerTask('i18next', 'Internationalization init', function() {
+  grunt.registerTask('i18n', 'Internationalization init', function() {
     var done = this.async();
     var options = this.options({
       preload: ['en'],
@@ -12,7 +12,7 @@ module.exports = function(grunt) {
         namespaces: ['main'],
         defaultNs: 'main'
       },
-      resGetPath: 'input/i18n/__lng__/__ns__.yaml'
+      resGetPath: 'input/i18n/__lng__/__ns__.yml'
     });
     Handlebars = require('handlebars');
