+++ tasks/options/clean.js
@@ -1,0 +1,29 @@
+module.exports = {
+
+  // 'grunt-contrib-clean': remove files
+
+  output: [
+    'output/'
+  ],
+
+  images: [
+    'output/assets/img/'
+  ],
+
+  assets: [
+    'output/assets/css/',
+    'output/assets/js/',
+    'output/assets/fonts/'
+  ],
+
+  templates: [
+    'output/templates/',
+    'output/i18n/',
+    'output/*.html'
+  ],
+
+  dist: [
+    '*.jar'
+  ]
+
+}
