+++ tasks/options/compile-handlebars.js
@@ -1,0 +1,20 @@
+module.exports = {
+
+  // 'grunt-compile-handlebars': converts Handlebars files and JSON into HTML
+
+  dist: {
+    files: [
+      {
+        expand: true,
+        cwd: 'output/templates',
+        src: '*.hbs',
+        dest: "output/<%= lng %>/",
+        ext: '.html'
+      }
+    ],
+    templateData: '*.json', // compile-handlebars uses the template folder no matter what
+    partials: 'input/templates/partials/**/*.hbs',
+    helpers: 'input/templates/helpers/**/*.js'
+  }
+
+}
