+++ tasks/options/imagemin.js
@@ -1,0 +1,14 @@
+module.exports = {
+
+  // 'grunt-contrib-imagemin': compresses images
+
+  dist: {
+    files: [{
+      expand: true,
+      cwd: 'output/assets/img/',
+      src: ['**/*.{png,jpg,gif,svg}'],
+      dest: 'output/assets/img/'
+    }]
+  }
+
+}
