+++ tasks/options/json-refs.js
@@ -1,0 +1,15 @@
+module.exports = {
+
+  // 'json-refs': resolves JSON references
+
+  dist: {
+    files: [{
+      expand: true,
+      cwd: 'input/templates',
+      src: '*.json',
+      dest: 'output/templates',
+      ext: '.json'
+    }]
+  }
+
+}
