+++ tasks/options/maven.js
@@ -1,0 +1,39 @@
+module.exports = {
+
+  // 'grunt-maven-tasks': generates the webjar and optionally places it to somewhere
+
+  options: {
+    type: "jar",
+    groupId: 'io.commercetools',
+    artifactId: "<%= pkg.name %>",
+    version: "<%= pkg.version %>",
+    destFolder: "/META-INF/resources/webjars",
+    gitpush: true,
+    gitpushtag: true
+  },
+
+  install: {
+    options: {
+      goal: "install"
+    },
+    files: [
+      { expand: true, cwd: 'output/assets/', src: "**/*", filter: "isFile" },
+      { expand: true, cwd: 'output/', src: "templates/**/*", filter: "isFile" },
+      { expand: true, cwd: 'output/', src: "i18n/**/*", filter: "isFile" }
+    ],
+  },
+
+  release: {
+    options: {
+      goal: "release",
+      repositoryId: "commercetools-bintray",
+      url: "https://api.bintray.com/maven/commercetools/maven/<%= pkg.name %>"
+    },
+    files: [
+      { expand: true, cwd: 'output/assets/', src: "**/*", filter: "isFile" },
+      { expand: true, cwd: 'output/', src: "templates/**/*", filter: "isFile" },
+      { expand: true, cwd: 'output/', src: "i18n/**/*", filter: "isFile" }
+    ],
+  }
+
+}
